Autenticacion creadda y aplicada a endpoints PUT, POST y DELETE de libros

+++ app/controllers/LibrosController.php
@@ -37,44 +37,53 @@ class LibrosController {
         if(empty($req)){
             return $this->getAll($req);
         }else{
-            $campos = ["isbn", "titulo", "fecha_de_publicacion", "editorial", "encuadernado", "sinopsis", "autor", "nro_de_paginas"];
-            $libro = new stdClass();
-            $isValidLibro = true;
-            foreach($campos as $campo){
-                $libro->$campo = $req->body->$campo ?? false;
-                $isValidLibro = $isValidLibro && $libro->$campo;
-            }
-            if($isValidLibro){
-                $this->model->create($libro);
-                return $this->view->response("Libro creado exitosamente".$descripcion,200);
-            }else{
-                $this->view->response("No se pude crear el libro porque hay campos vacios.", 404);
-            }
-        }
-    }
-    public function update($req){
-        if(empty($req)){
-            return $this->getAll($req);
-        }else{
-            $id=$req->params->ID;
-            $libro = $this->model->find($id);
-            if($libro){
-                $campos = ["isbn", "titulo", "fecha_de_publicacion", "editorial", "encuadernado", "sinopsis", "autor", "nro_de_paginas", "old_isbn"];
+            if(AuthHelper::verifyJWTToken($req)){
+                $campos = ["isbn", "titulo", "fecha_de_publicacion", "editorial", "encuadernado", "sinopsis", "autor", "nro_de_paginas"];
                 $libro = new stdClass();
                 $isValidLibro = true;
                 foreach($campos as $campo){
                     $libro->$campo = $req->body->$campo ?? false;
                     $isValidLibro = $isValidLibro && $libro->$campo;
                 }
-
                 if($isValidLibro){
-                    $this->model->update($libro);
-                    $this->view->response("El libro con id ".$id." se actualizo de forma correcta.", 200);
+                    $this->model->create($libro);
+                    return $this->view->response("Libro creado exitosamente".$descripcion,200);
                 }else{
-                    $this->view->response("El libro no se puede actualizar porque hay campos vacíos.", 404);
+                    $this->view->response("No se pude crear el libro porque hay campos vacios.", 404);
                 }
             }else{
-                $this->view->response("El libro no existe.", 404);
+                $this->view->response("No tienes permiso para acceder a este recurso", 401);
+            }
+
+        }
+    }
+    public function update($req){
+        if(empty($req)){
+            return $this->getAll($req);
+        }else{
+            if(AuthHelper::verifyJWTToken($req)){
+                $id=$req->params->ID;
+                $libro = $this->model->find($id);
+                if($libro){
+                    $campos = ["isbn", "titulo", "fecha_de_publicacion", "editorial", "encuadernado", "sinopsis", "autor", "nro_de_paginas", "old_isbn"];
+                    $libro = new stdClass();
+                    $isValidLibro = true;
+                    foreach($campos as $campo){
+                        $libro->$campo = $req->body->$campo ?? false;
+                        $isValidLibro = $isValidLibro && $libro->$campo;
+                    }
+
+                    if($isValidLibro){
+                        $this->model->update($libro);
+                        $this->view->response("El libro con id ".$id." se actualizo de forma correcta.", 200);
+                    }else{
+                        $this->view->response("El libro no se puede actualizar porque hay campos vacíos.", 404);
+                    }
+                }else{
+                    $this->view->response("El libro no existe.", 404);
+                }
+            }else{
+                $this->view->response("No tienes permiso para acceder a este recurso", 401);
             }
         }
     }
@@ -82,12 +91,16 @@ class LibrosController {
         if(empty($req->params->ID)){
             return $this->getAll($req);
         }else{
-            $libro = $this->model->find($req->params->ID);
-            if($libro){
-                $this->model->delete($req->params->ID);
-                $this->view->response("El libro ".$libro->titulo." se elimino correctamente.", 200);
+            if(AuthHelper::verifyJWTToken($req)){
+                $libro = $this->model->find($req->params->ID);
+                if($libro){
+                    $this->model->delete($req->params->ID);
+                    $this->view->response("El libro ".$libro->titulo." se elimino correctamente.", 200);
+                }else{
+                    $this->view->response("El libro con la id ".$req->params->ID." no se encuentra en registrado.", 404);
+                }
             }else{
-                $this->view->response("El libro con la id ".$req->params->ID." no se encuentra en registrado.", 404);
+                $this->view->response("No tienes permiso para acceder a este recurso", 401);
             }
         }
     }
